requirements.txt
# Core dependencies for geospatial analysis
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
rasterio>=1.3.0
xarray>=2023.1.0
netcdf4>=1.6.0
h5netcdf>=1.2.0

# Geospatial processing (GDAL/GEOS dependencies)
Fiona>=1.9.0
Shapely>=2.0.0
pyproj>=3.6.0
rasterstats>=0.18.0

# Climate and earth system data
dask>=2023.1.0
zarr>=2.14.0
cartopy>=0.21.0
cf-xarray>=0.8.0

# Visualization and web
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0
altair>=5.0.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Data acquisition and APIs
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
jupyterlab>=4.0.0

# Development and testing
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.0.0

# Atlas data processing and validation  
tqdm>=4.65.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Enhanced geospatial and web preparation
contextily>=1.4.0
geojson>=3.0.0
topojson>=1.5.0

# Optional: Advanced geospatial processing
# Uncomment if needed for advanced features
# pygeos>=0.14.0
# rtree>=1.0.0
# earthengine-api>=0.1.360

# Observable Framework preparation (run separately with npm)
# npm create @observablehq/framework@latest soil-health-ssa

# Windows-specific GDAL installation note:
# If you encounter GDAL installation issues on Windows:
# 1. Download GDAL wheel from https://www.lfd.uci.edu/~gohlke/pythonlibs/#gdal
# 2. Install with: pip install path/to/GDAL-X.X.X-cpXX-cpXX-win_amd64.whl
# 3. Or use OSGeo4W: https://trac.osgeo.org/osgeo4w/

# Current Status: Atlas data pipeline complete ✅
# Next: SoilGrids integration → Observable Framework deployment